updated to the latest bmd code, checked that docker based running with these latest code work fine
