+++ code/latest/1_tall2wide/3_after_tall2wide/count/count_empty_responses/working_example/input/344_zf_morphology_data_phase_1_and_2_-_2020JUNE25_wide_DNC_0_head.csv
@@ -1,0 +1,3 @@
+chemical.id,plate.id,well,chemical_plate_well,conc,AXIS,BRAI,CFIN,CIRC,DNC_,DP24,EYE_,JAW_,MO24,MORT,NC24,NC__,OTIC,PE__,PFIN,PIG_,SM24,SNOU,SOMI,SWIM,TRUN,TR__,YSE_
+53,9289,A01,53_9289_A01,50,,,,,0,0,,,0,1,0,,,,,,0,,,,,,
+53,9289,A02,53_9289_A02,35.6,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0
