+++ code/latest/1_tall2wide/3_after_tall2wide/investigate_NaNs_w_MO24_MORT/working_example/output/has_at_least_1_non_nan_response_other_than_24series_and_MORT_even_when_MORT_eq_1.csv
@@ -1,0 +1,2 @@
+chemical.id,plate.id,well,chemical_plate_well,conc,AXIS,BRAI,CFIN,CIRC,DNC_,DP24,EYE_,JAW_,MO24,MORT,NC24,NC__,OTIC,PE__,PFIN,PIG_,SM24,SNOU,SOMI,SWIM,TRUN,TR__,YSE_
+1853,2,H04,1853_2_H04,0.064,1.0,1.0,1.0,1.0,0.0,,1.0,1.0,1.0,1.0,,1.0,1.0,1.0,1.0,1.0,,1.0,1.0,,1.0,1.0,1.0
