+++ code/latest/2_bmd/common/count_0_responses/working_example/input/dose_response_vals_qc_1_new_head.csv
@@ -1,0 +1,19 @@
+Chemical_ID,End_Point,Dose,Response,CI_Lo,CI_Hi
+53,ANY24,0.0,0.09375,0.061348448701633834,0.14843154728387867
+53,ANY24,1.0,0.0625,0.04518943296835905,0.13897123144154583
+53,ANY24,5.0,0.125,0.07529865612519929,0.15568305481650224
+53,ANY24,11.2,0.0625,0.04518943296835905,0.13897123144154583
+53,ANY24,35.6,0.09375,0.061348448701633834,0.14843154728387867
+53,ANY24,50.0,0.03125,0.025712139835996878,0.12619263820012558
+53,ANY120,0.0,0.125,0.07529865612519929,0.15568305481650224
+53,ANY120,1.0,0.0625,0.04518943296835905,0.13897123144154583
+53,ANY120,5.0,0.15625,0.08760579717494293,0.16129149597530396
+53,ANY120,11.2,0.09375,0.061348448701633834,0.14843154728387867
+53,ANY120,35.6,0.09375,0.061348448701633834,0.14843154728387867
+53,ANY120,50.0,0.09375,0.061348448701633834,0.14843154728387867
+53,DP24,0.0,0.0,6.938893903907228e-18,0.11696979849974073
+53,DP24,1.0,0.0625,0.04518943296835905,0.13897123144154583
+53,DP24,5.0,0.0967741935483871,0.0633095493851897,0.15222519088554123
+53,DP24,11.2,0.03225806451612903,0.026540842990712216,0.12968298713112003
+53,DP24,35.6,0.09375,0.061348448701633834,0.14843154728387867
+53,DP24,50.0,0.03125,0.025712139835996878,0.12619263820012558
