+++ code/latest/2_bmd/common/count_0_responses/working_example/output/Dose_count_zero_Response.csv
@@ -1,0 +1,4 @@
+Dose_count,zero_Response,one_response,zero_Response/Dose_count
+6,0,6,0.0%
+6,0,6,0.0%
+6,1,5,17.0%
