+++ code/paritosh_original/qc_BMD/behavior_phase_I_II/3_bmd_feasibility_BMD/working/bmd_vals.csv
@@ -1,0 +1,26 @@
+Chemical_ID,End_Point,Model,BMD10,BMDL,BMD50,AUC,Min_Dose,Max_Dose,AUC_Norm,DataQC_Flag,BMD_Analysis_Flag,BMD10_Flag,BMD50_Flag
+53,MOV1,weibull,0.0,NULL,584003758.644,16.159605414746544,0.0,50.0,0.3231921082949309,4,0,-1,1
+53,AUC1,logistic,73.726,29.709,185.957,6.276771313364055,0.0,50.0,0.1255354262672811,4,0,1,1
+54,MOV1,log_probit,0.0,NULL,51168.921,1.3256107276236309,0.0,5.0,0.2651221455247262,4,0,-1,1
+54,AUC1,log_probit,0.0,NULL,311097.287,1.2032107276236308,0.0,5.0,0.24064214552472615,4,0,-1,1
+53,MOV1,weibull,0.0,NULL,584003758.644,16.159605414746544,0.0,50.0,0.3231921082949309,4,0,-1,1
+53,AUC1,logistic,73.726,29.709,185.957,6.276771313364055,0.0,50.0,0.1255354262672811,4,0,1,1
+54,MOV1,log_probit,0.0,NULL,51168.921,1.3256107276236309,0.0,5.0,0.2651221455247262,4,0,-1,1
+54,AUC1,log_probit,0.0,NULL,311097.287,1.2032107276236308,0.0,5.0,0.24064214552472615,4,0,-1,1
+53,MOV1,weibull,0.0,NULL,584003758.644,16.159605414746544,0.0,50.0,0.3231921082949309,4,0,-1,1
+53,AUC1,logistic,73.726,29.709,185.957,6.276771313364055,0.0,50.0,0.1255354262672811,4,0,1,1
+54,MOV1,log_probit,0.0,NULL,51168.921,1.3256107276236309,0.0,5.0,0.2651221455247262,4,0,-1,1
+54,AUC1,log_probit,0.0,NULL,311097.287,1.2032107276236308,0.0,5.0,0.24064214552472615,4,0,-1,1
+53,MOV1,weibull,0.0,NULL,584003758.644,16.159605414746544,0.0,50.0,0.3231921082949309,4,0,-1,1
+53,AUC1,logistic,73.726,29.709,185.957,6.276771313364055,0.0,50.0,0.1255354262672811,4,0,1,1
+54,MOV1,log_probit,0.0,NULL,51168.921,1.3256107276236309,0.0,5.0,0.2651221455247262,4,0,-1,1
+54,AUC1,log_probit,0.0,NULL,311097.287,1.2032107276236308,0.0,5.0,0.24064214552472615,4,0,-1,1
+54,MOV1,log_probit,0.0,NULL,51168.921,1.3256107276236309,0.0,5.0,0.2651221455247262,4,0,-1,1
+54,AUC1,log_probit,0.0,NULL,311097.287,1.2032107276236308,0.0,5.0,0.24064214552472615,4,0,-1,1
+54,MOV1,log_probit,0.0,NULL,51168.921,1.3256107276236309,0.0,5.0,0.2651221455247262,4,0,-1,1
+54,AUC1,log_probit,0.0,NULL,311097.287,1.2032107276236308,0.0,5.0,0.24064214552472615,4,0,-1,1
+54,MOV1,log_probit,0.0,NULL,51168.921,1.3256107276236309,0.0,5.0,0.2651221455247262,4,0,-1,1
+54,MOV1,log_probit,0.0,NULL,51168.921,1.3256107276236309,0.0,5.0,0.2651221455247262,4,0,-1,1
+54,AUC1,log_probit,0.0,NULL,311097.287,1.2032107276236308,0.0,5.0,0.24064214552472615,4,0,-1,1
+54,MOV1,log_probit,0.0,NULL,51168.921,1.3256107276236309,0.0,5.0,0.2651221455247262,4,0,-1,1
+54,AUC1,log_probit,0.0,NULL,311097.287,1.2032107276236308,0.0,5.0,0.24064214552472615,4,0,-1,1
