+++ code/paritosh_original/qc_BMD/behavior_phase_I_II/3_bmd_feasibility_BMD/working/output/dose_response_vals.csv
@@ -1,0 +1,13 @@
+Chemical_ID,End_Point,Dose,Response,CI_Lo,CI_Hi
+54,MOV1,0.0,0.038461538461538464,0.031639553467995005,0.1504662505688091
+54,MOV1,0.1,0.3,0.133352517317562,0.17875787458714948
+54,MOV1,0.5,0.25806451612903225,0.12104745706509479,0.1743968419653057
+54,MOV1,1.12,0.32,0.14794809693092012,0.19589730726897753
+54,MOV1,3.56,0.2222222222222222,0.13221294113620524,0.22992362093990398
+54,MOV1,5.0,0.2727272727272727,0.17526667982702376,0.2929180284488564
+54,AUC1,0.0,0.038461538461538464,0.031639553467995005,0.1504662505688091
+54,AUC1,0.1,0.3,0.133352517317562,0.17875787458714948
+54,AUC1,0.5,0.25806451612903225,0.12104745706509479,0.1743968419653057
+54,AUC1,1.12,0.24,0.12503686306598552,0.19429683355429073
+54,AUC1,3.56,0.2222222222222222,0.13221294113620524,0.22992362093990398
+54,AUC1,5.0,0.2727272727272727,0.17526667982702376,0.2929180284488564
