+++ code/paritosh_original/qc_BMD/behavior_phase_I_II/3_bmd_feasibility_BMD/working/output/fit_vals.csv
@@ -1,0 +1,93 @@
+Chemical_ID,End_Point,X_vals,Y_vals
+54,MOV1,0.0,0.04726919842234318
+54,MOV1,0.011111111111111112,0.2082069151777474
+54,MOV1,0.022222222222222223,0.21889088342758156
+54,MOV1,0.03333333333333333,0.22528410415383435
+54,MOV1,0.044444444444444446,0.2298835799477415
+54,MOV1,0.05555555555555556,0.2334871180662602
+54,MOV1,0.06666666666666667,0.2364545325628328
+54,MOV1,0.07777777777777778,0.23897955468479373
+54,MOV1,0.08888888888888889,0.24117869139251405
+54,MOV1,0.1,0.24312756231560934
+54,MOV1,0.14444444444444446,0.24926645244569018
+54,MOV1,0.18888888888888888,0.25379638433710516
+54,MOV1,0.23333333333333334,0.2573947674618472
+54,MOV1,0.2777777777777778,0.2603836703925115
+54,MOV1,0.32222222222222224,0.2629420178436869
+54,MOV1,0.3666666666666667,0.26517968560475946
+54,MOV1,0.4111111111111111,0.2671690664287282
+54,MOV1,0.4555555555555556,0.2689604190865476
+54,MOV1,0.5,0.2705900782072593
+54,MOV1,0.5688888888888889,0.27285785739840224
+54,MOV1,0.6377777777777778,0.2748739394790067
+54,MOV1,0.7066666666666668,0.2766892597310181
+54,MOV1,0.7755555555555556,0.2783406475183463
+54,MOV1,0.8444444444444446,0.2798556064819543
+54,MOV1,0.9133333333333334,0.2812552250730114
+54,MOV1,0.9822222222222223,0.28255603218833697
+54,MOV1,1.051111111111111,0.28377122616452444
+54,MOV1,1.12,0.2849115152528906
+54,MOV1,1.3911111111111112,0.288821755321405
+54,MOV1,1.6622222222222223,0.2920523021324897
+54,MOV1,1.9333333333333333,0.29480677298179087
+54,MOV1,2.2044444444444444,0.2972088517562392
+54,MOV1,2.4755555555555553,0.2993393796558326
+54,MOV1,2.7466666666666666,0.3012541365118556
+54,MOV1,3.017777777777778,0.3029932512600501
+54,MOV1,3.2888888888888888,0.30458656839270276
+54,MOV1,3.56,0.3060568928412578
+54,MOV1,3.72,0.3068742968485785
+54,MOV1,3.88,0.30765817495077
+54,MOV1,4.04,0.3084111990225943
+54,MOV1,4.2,0.30913573250822496
+54,MOV1,4.36,0.30983387623162423
+54,MOV1,4.52,0.3105075059955776
+54,MOV1,4.68,0.31115830367600134
+54,MOV1,4.84,0.31178778311792155
+54,MOV1,5.0,0.31239731184313935
+54,AUC1,0.0,0.048446634626929636
+54,AUC1,0.011111111111111112,0.1968128374853014
+54,AUC1,0.022222222222222223,0.2065136065454537
+54,AUC1,0.03333333333333333,0.21231718749654283
+54,AUC1,0.044444444444444446,0.2164920888041077
+54,AUC1,0.05555555555555556,0.21976285744130292
+54,AUC1,0.06666666666666667,0.2224562025767652
+54,AUC1,0.07777777777777778,0.22474800805703318
+54,AUC1,0.08888888888888889,0.22674403766337298
+54,AUC1,0.1,0.2285129331508946
+54,AUC1,0.14444444444444446,0.234085106004945
+54,AUC1,0.18888888888888888,0.238197155032074
+54,AUC1,0.23333333333333334,0.2414638440029463
+54,AUC1,0.2777777777777778,0.24417744126401064
+54,AUC1,0.32222222222222224,0.24650031186438615
+54,AUC1,0.3666666666666667,0.2485321614154442
+54,AUC1,0.4111111111111111,0.25033868184446334
+54,AUC1,0.4555555555555556,0.25196548005759206
+54,AUC1,0.5,0.25344552788073643
+54,AUC1,0.5688888888888889,0.2555052632277177
+54,AUC1,0.6377777777777778,0.25733654646515175
+54,AUC1,0.7066666666666668,0.2589856010140818
+54,AUC1,0.7755555555555556,0.2604858500812157
+54,AUC1,0.8444444444444446,0.2618622542688181
+54,AUC1,0.9133333333333334,0.2631339526438614
+54,AUC1,0.9822222222222223,0.26431594658769325
+54,AUC1,1.051111111111111,0.2654202150266435
+54,AUC1,1.12,0.2664564771156789
+54,AUC1,1.3911111111111112,0.2700104579263059
+54,AUC1,1.6622222222222223,0.2729472437327149
+54,AUC1,1.9333333333333333,0.27545167562507267
+54,AUC1,2.2044444444444444,0.27763604244981194
+54,AUC1,2.4755555555555553,0.2795737437304467
+54,AUC1,2.7466666666666666,0.2813154286397969
+54,AUC1,3.017777777777778,0.28289753777105725
+54,AUC1,3.2888888888888888,0.28434717446546603
+54,AUC1,3.56,0.2856850504666545
+54,AUC1,3.72,0.28642888125912924
+54,AUC1,3.88,0.2871422441601026
+54,AUC1,4.04,0.2878275661667936
+54,AUC1,4.2,0.28848699418679197
+54,AUC1,4.36,0.2891224366320578
+54,AUC1,4.52,0.28973559755822514
+54,AUC1,4.68,0.2903280048984289
+54,AUC1,4.84,0.2909010339775973
+54,AUC1,5.0,0.29145592722411084
