+++ qc_BMD/BMD_Analysis_Functions.py
@@ -10,8 +10,14 @@ import numpy as np
 from scipy import stats
 from statsmodels.base.model import GenericLikelihoodModel
 
+import warnings
+warnings.filterwarnings('ignore')
+
 BMR = "global"
 BMR = 0.1
+
+#report = True
+report = False
 
 # Logistic function
 
@@ -47,7 +53,7 @@ class Logistic(GenericLikelihoodModel):
             alpha_0 = -mu_0/s_0
             beta_0 = 1/s_0
             start_params = np.array([alpha_0, beta_0])
-            print('start_params [alpha, beta]: ' + str(start_params))
+            #print('start_params [alpha, beta]: ' + str(start_params))
             
         return super(Logistic, self).fit(start_params = start_params, maxiter = maxiter, maxfun = maxfun, method = 'lbfgs', bounds = [[None, None],[1e-5,None]], disp=0, **kwds)
       
@@ -111,7 +117,8 @@ class Gamma(GenericLikelihoodModel):
             beta_0 = self.endog[:,0].flatten().mean()/self.endog[:,0].flatten().var()
             alpha_0 = self.endog[:,0].flatten().mean() * beta_0
             start_params = np.array([g_0, alpha_0, beta_0])
-            print('start_params [g, alpha, beta]: ' + str(start_params))
+            if (report):
+                print('start_params [g, alpha, beta]: ' + str(start_params))
 
         return super(Gamma, self).fit(start_params = start_params, maxiter = maxiter, maxfun = maxfun, method = 'lbfgs', bounds = [[1e-5,0.99],[0.2, 18],[1e-5, None]], **kwds)
         
@@ -186,7 +193,8 @@ class Weibull(GenericLikelihoodModel):
             beta_0 = np.exp(betas[0])
             
             start_params = np.array([g_0, alpha_0, beta_0])
-            print('start_params [g, alpha, beta]: ' + str(start_params))
+            if (report):
+                print('start_params [g, alpha, beta]: ' + str(start_params))
             
         return super(Weibull, self).fit(start_params = start_params, maxiter = maxiter, maxfun = maxfun, method = 'lbfgs', bounds = [[1e-5,0.99],[1e-5,None],[1e-9,None]], **kwds)
     
@@ -254,7 +262,8 @@ class Log_Logistic(GenericLikelihoodModel):
             alpha_0 = -mu_0/s_0
             beta_0 = 1/s_0
             start_params = np.array([g_0, alpha_0, beta_0])
-            print('start_params [g, alpha, beta]: ' + str(start_params))
+            if (report):
+                print('start_params [g, alpha, beta]: ' + str(start_params))
             
         return super(Log_Logistic, self).fit(start_params = start_params, maxiter = maxiter, maxfun = maxfun, method = 'lbfgs', bounds = [[1e-5,0.99],[None, None],[None, None]], **kwds)
     
@@ -317,7 +326,8 @@ class Probit(GenericLikelihoodModel):
             alpha_0 = stats.norm.ppf(num_affected[0]/num_total[0])
             beta_0 = (stats.norm.ppf(num_affected[-1]/num_total[-1]) - alpha_0)/dose[-1] 
             start_params = np.array([alpha_0, beta_0])
-            print('start_params [alpha, beta]: ' + str(start_params))
+            if (report):
+                print('start_params [alpha, beta]: ' + str(start_params))
 
         return super(Probit, self).fit(start_params = start_params, maxiter = maxiter, maxfun = maxfun, method = 'lbfgs', bounds = [[None, None],[1e-5,None]], **kwds)
     
@@ -396,7 +406,8 @@ class Log_Probit(GenericLikelihoodModel):
             
             dose = self.endog[:,0].flatten()
             start_params = np.array([g_0, alpha_0, beta_0])
-            print('start_params [g, alpha, beta]: ' + str(start_params))
+            if (report):
+                print('start_params [g, alpha, beta]: ' + str(start_params))
             
         return super(Log_Probit, self).fit(start_params = start_params, maxiter = maxiter, maxfun = maxfun, method = 'lbfgs', bounds = [[1e-5,0.99],[None,None],[1e-5,None]], **kwds)
     
@@ -471,7 +482,8 @@ class Multistage_2(GenericLikelihoodModel):
             beta2_0 = betas[1]
 
             start_params = np.array([g_0, beta1_0, beta2_0])
-            print('start_params [g, beta 1, beta 2]: ' + str(start_params))
+            if (report):
+                print('start_params [g, beta 1, beta 2]: ' + str(start_params))
             
         return super(Multistage_2, self).fit(start_params = start_params, maxiter = maxiter, maxfun = maxfun, method = 'lbfgs', bounds = [[1e-9,0.99],[1e-9,None],[1e-9,None]], **kwds)
         
@@ -531,7 +543,8 @@ class Quantal_Linear(GenericLikelihoodModel):
             g_0 = 0.1
             beta_0 = 1/((self.endog[:,0].flatten().mean()))/np.log(2)
             start_params = np.array([g_0, beta_0])
-            print('start_params [g, beta]: ' + str(start_params))
+            if (report):
+                print('start_params [g, beta]: ' + str(start_params))
             
         return super(Quantal_Linear, self).fit(start_params = start_params, maxiter = maxiter, maxfun = maxfun, method = 'lbfgs', bounds = [[1e-5,0.99],[1e-5,None]], **kwds)
